added the pots command to display the list of pots. exported the data in json files to be loaded at startup.

+++ sources/CaveOfWonders.Cli.Application/CurrencyValue.cs
@@ -18,7 +18,7 @@ namespace DustInTheWind.CaveOfWonders.Cli.Application;
 
 public class CurrencyValue
 {
-    public float Value { get; set; }
+    public double Value { get; set; }
 
     public string Currency { get; set; }
 }
