+++ sources/CaveOfWonders.Domain/Gem.cs
@@ -20,5 +20,5 @@ public class Gem
 {
     public DateTime Date { get; set; }
 
-    public float Value { get; set; }
+    public double Value { get; set; }
 }
